2009-01-16  Tatsuhiro Tsujikawa  <t-tujikawa@users.sourceforge.net>
	Fixed compile warning/error.
	* src/DownloadResult.h
	* src/HttpDownloadCommand.cc
	* src/ServerStat.cc
	* test/RequestGroupTest.cc

+++ src/HttpDownloadCommand.cc
@@ -93,7 +93,7 @@ bool HttpDownloadCommand::prepareForNextSegment() {
     if(!req->isPipeliningEnabled() && req->isPipeliningHint() &&
        !_segments.empty() && !downloadFinished) {
       const SharedHandle<Segment>& segment = _segments.front();
-      if(segment->getPosition()+segment->getLength() ==
+      if(static_cast<uint64_t>(segment->getPosition())+segment->getLength() ==
 	 static_cast<uint64_t>(_httpResponse->getHttpHeader()->
 			       getRange()->getEndByte()+1)) {
 	return prepareForRetry(0);
